requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
langchain 
langgraph 
langchain-community
streamlit
python-dotenv
langchain-google-genai